+++ daily-schedule-be/src/main/java/daily_schedule/daily_schedule_be/domain/todo/entity/ScheduleResult.java
@@ -1,4 +1,4 @@
-package daily_schedule.daily_schedule_be.domain;
+package daily_schedule.daily_schedule_be.domain.todo.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
