+++ daily-schedule-be/src/main/java/daily_schedule/daily_schedule_be/domain/todo/repository/SchedulesResultRepository.java
@@ -1,6 +1,6 @@
-package daily_schedule.daily_schedule_be.repository;
+package daily_schedule.daily_schedule_be.domain.todo.repository;
 
-import daily_schedule.daily_schedule_be.domain.ScheduleResult;
+import daily_schedule.daily_schedule_be.domain.todo.entity.ScheduleResult;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SchedulesResultRepository extends JpaRepository<ScheduleResult, Long>{
